Typescriptify Clients-Component
