+++ ui/src/pages/dialog/AddClientDialog.tsx
@@ -1,0 +1,49 @@
+import Button from 'material-ui/Button';
+import Dialog, {DialogActions, DialogContent, DialogTitle} from 'material-ui/Dialog';
+import TextField from 'material-ui/TextField';
+import Tooltip from 'material-ui/Tooltip';
+import React, {Component} from 'react';
+
+interface IProps {
+    fClose: VoidFunction
+    fOnSubmit: (name: string) => void
+}
+
+export default class AddDialog extends Component<IProps, { name: string }> {
+    public state = {name: ''};
+
+    public render() {
+        const {fClose, fOnSubmit} = this.props;
+        const {name} = this.state;
+        const submitEnabled = this.state.name.length !== 0;
+        const submitAndClose = () => {
+            fOnSubmit(name);
+            fClose();
+        };
+        return (
+            <Dialog open={true} onClose={fClose} aria-labelledby="form-dialog-title">
+                <DialogTitle id="form-dialog-title">Create a client</DialogTitle>
+                <DialogContent>
+                    <TextField autoFocus margin="dense" id="name" label="Name *" type="email" value={name}
+                               onChange={this.handleChange.bind(this, 'name')} fullWidth/>
+                </DialogContent>
+                <DialogActions>
+                    <Button onClick={fClose}>Cancel</Button>
+                    <Tooltip placement={'bottom-start'} title={submitEnabled ? '' : 'name is required'}>
+                        <div>
+                            <Button disabled={!submitEnabled} onClick={submitAndClose} color="primary" variant="raised">
+                                Create
+                            </Button>
+                        </div>
+                    </Tooltip>
+                </DialogActions>
+            </Dialog>
+        );
+    }
+
+    private handleChange(propertyName: string, event: React.ChangeEvent<HTMLInputElement>) {
+        const state = this.state;
+        state[propertyName] = event.target.value;
+        this.setState(state);
+    }
+}
